+++ crates/parry2d/tests/geometry/ball_ball_toi.rs
@@ -1,8 +1,8 @@
 // Issue #35
 
 use na::{self, Isometry2, Vector2};
-use parry2d::math::Real;
 use parry2d::query;
+use parry2d::query::details::ShapeCastOptions;
 use parry2d::shape::Ball;
 
 #[test]
@@ -13,7 +13,7 @@ fn test_ball_ball_toi() {
     let v1 = Vector2::new(0.0, 10.0);
     let v2 = Vector2::zeros();
 
-    let cast = query::time_of_impact(&m1, &v1, &b, &m2, &v2, &b, Real::MAX, true).unwrap();
+    let cast = query::cast_shapes(&m1, &v1, &b, &m2, &v2, &b, ShapeCastOptions::default()).unwrap();
 
-    assert_eq!(cast.unwrap().toi, 0.9);
+    assert_eq!(cast.unwrap().time_of_impact, 0.9);
 }