+++ crates/parry3d/tests/geometry/time_of_impact3.rs
@@ -1,6 +1,6 @@
 use na::{self, Isometry3, Vector3};
 use parry3d::math::Real;
-use parry3d::query;
+use parry3d::query::{self, ShapeCastOptions};
 use parry3d::shape::{Ball, Cuboid};
 
 #[test]
@@ -19,42 +19,39 @@ fn ball_cuboid_toi() {
     let ball_vel1 = Vector3::new(2.0, 2.0, 2.0);
     let ball_vel2 = Vector3::new(-0.5, -0.5, -0.5);
 
-    let toi_intersecting = query::time_of_impact(
+    let toi_intersecting = query::cast_shapes(
         &ball_pos_intersecting,
         &ball_vel1,
         &ball,
         &cuboid_pos,
         &cuboid_vel1,
         &cuboid,
-        Real::MAX,
-        true,
+        ShapeCastOptions::default(),
     )
     .unwrap()
-    .map(|toi| toi.toi);
-    let toi_will_touch = query::time_of_impact(
+    .map(|hit| hit.time_of_impact);
+    let toi_will_touch = query::cast_shapes(
         &ball_pos_will_touch,
         &ball_vel2,
         &ball,
         &cuboid_pos,
         &cuboid_vel2,
         &cuboid,
-        Real::MAX,
-        true,
+        ShapeCastOptions::default(),
     )
     .unwrap()
-    .map(|toi| toi.toi);
-    let toi_wont_touch = query::time_of_impact(
+    .map(|hit| hit.time_of_impact);
+    let toi_wont_touch = query::cast_shapes(
         &ball_pos_wont_touch,
         &ball_vel1,
         &ball,
         &cuboid_pos,
         &cuboid_vel1,
         &cuboid,
-        Real::MAX,
-        true,
+        ShapeCastOptions::default(),
     )
     .unwrap()
-    .map(|toi| toi.toi);
+    .map(|hit| hit.time_of_impact);
 
     assert_eq!(toi_intersecting, Some(0.0));
     assert!(relative_eq!(