+++ src/query/visitors/ray_intersections_visitor.rs
@@ -8,7 +8,7 @@ use std::marker::PhantomData;
 /// Bounding Volume Tree visitor collecting intersections with a given ray.
 pub struct RayIntersectionsVisitor<'a, T, F> {
     simd_ray: SimdRay,
-    max_toi: SimdReal,
+    max_time_of_impact: SimdReal,
     callback: &'a mut F,
     _phantom: PhantomData<T>,
 }
@@ -19,10 +19,14 @@ where
 {
     /// Creates a new `RayIntersectionsVisitor`.
     #[inline]
-    pub fn new(ray: &Ray, max_toi: Real, callback: &'a mut F) -> RayIntersectionsVisitor<'a, T, F> {
+    pub fn new(
+        ray: &Ray,
+        max_time_of_impact: Real,
+        callback: &'a mut F,
+    ) -> RayIntersectionsVisitor<'a, T, F> {
         RayIntersectionsVisitor {
             simd_ray: SimdRay::splat(*ray),
-            max_toi: SimdReal::splat(max_toi),
+            max_time_of_impact: SimdReal::splat(max_time_of_impact),
             callback,
             _phantom: PhantomData,
         }
@@ -35,7 +39,7 @@ where
 {
     #[inline]
     fn visit(&mut self, bv: &SimdAabb, b: Option<[Option<&T>; SIMD_WIDTH]>) -> SimdVisitStatus {
-        let mask = bv.cast_local_ray(&self.simd_ray, self.max_toi).0;
+        let mask = bv.cast_local_ray(&self.simd_ray, self.max_time_of_impact).0;
 
         if let Some(data) = b {
             let bitmask = mask.bitmask();